auth/autorest: Move methods that require autorest to their own module
This makes it possible to use hamilton without importing everything from
the Microsoft autorest library, which is quite nice. :^)
