+++ auth/auth.go
@@ -8,7 +8,6 @@ import (
 	"io/ioutil"
 	"strings"
 
-	"github.com/Azure/go-autorest/autorest"
 	"golang.org/x/crypto/pkcs12"
 	"golang.org/x/oauth2"
 
@@ -80,11 +79,6 @@ func (c *Config) NewAuthorizer(ctx context.Context, api environments.Api) (Autho
 	}
 
 	return nil, fmt.Errorf("no Authorizer could be configured, please check your configuration")
-}
-
-// NewAutorestAuthorizerWrapper returns an Authorizer that sources tokens from a supplied autorest.BearerAuthorizer
-func NewAutorestAuthorizerWrapper(autorestAuthorizer autorest.Authorizer) (Authorizer, error) {
-	return &AutorestAuthorizerWrapper{authorizer: autorestAuthorizer}, nil
 }
 
 // NewAzureCliAuthorizer returns an Authorizer which authenticates using the Azure CLI.
