+++ auth/autorest/autorest.go
@@ -1,4 +1,4 @@
-package auth
+package autorest
 
 import (
 	"context"
